+++ modules/search-box/search-box.php
@@ -6,6 +6,7 @@
 		'value' => get_search_query() ?: get_query_var( 'q', false ),
 		'index' => false,
 		'suggestions' => false,
+		'facetFilters' => '',
 		'modifiers' => ['large']
 	];
 
@@ -13,11 +14,12 @@
 	$data->modifiers = telabotanica_styleguide_modifiers_array('search-box', $data->modifiers);
 
 	printf(
-		'<div class="%s" data-autocomplete="%s" data-instantsearch="%s" data-index="%s">',
+		'<div class="%s" data-autocomplete="%s" data-instantsearch="%s" data-index="%s" data-facet-filters="%s">',
 		implode(' ', $data->modifiers),
 		var_export($data->autocomplete, true),
 		var_export($data->instantsearch, true),
-		$data->index
+		$data->index,
+		$data->facetFilters
 	);
 		printf(
 			'<form role="search" method="get" action="%s" class="search-box-wrapper">',